requirements.txt
fastmcp>=2.14.5
